refactor: 목데이터를 consts에서 mocks 폴더로 이동
